# openapi.yml
openapi: 3.0.3
info:
  title: "Pulse service APIs"
  description: "The api to export the data bruh"
  version: 1.0.0
servers:
  - url: https://localhost/

paths:
  /exporters/web/uptime:
    get:
      summary: "Get uptime metrics"
      description: "If queried data range is unspecified, it defaults to the last 6 hours"
      parameters:
        - name: from
          in: query
          description: "Timeframe start point"
          required: false
          schema:
            type: string
        - name: to
          in: query
          description: "Timeframe start end"
          required: false
          schema:
            type: string
        - name: interval
          in: query
          description: "Series aggregation interval"
          required: false
          schema:
            type: string
      responses:
        '200':
          description: "Successful query"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: string
                      description: "Data point timestamp"
                      example: "2025-03-07T19:38:08.746+01:00"
                    label:
                      type: string
                      description: "Probe label"
                      example: "google.com"
                    status:
                      type: string
                      description: "Service status string"
                      example: "up"
                    http_status:
                      type: integer
                      description: "Http response status code"
                      example: 200
                    elapsed_ms:
                      type: integer
                      description: "Task execution duration (always non-zero)"
                      example: 150
                    latency_ms:
                      type: integer
                      description: "Service latency (>= 0 if up, otherwise == -1)"
                      example: 150
  /exporters/web/tlscert:
    get:
      summary: "Get TLS cert metrics"
      description: "If queried data range is unspecified, it defaults to the last one hour"
      parameters:
        - name: from
          in: query
          description: "Timeframe start point"
          required: false
          schema:
            type: string
        - name: to
          in: query
          description: "Timeframe start end"
          required: false
          schema:
            type: string
      responses:
        '200':
          description: "Successful query"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: string
                      description: "Data point timestamp"
                      example: "2025-03-07T19:38:08.746+01:00"
                    label:
                      type: string
                      description: "Probe label"
                      example: "google.com"
                    security:
                      type: string
                      description: "Service security status"
                      example: "tls 1.3"
                    cert_subject:
                      type: string
                      description: "Certificate issue subject"
                      example: "CN=*.google.com"
                    cert_issuer:
                      type: string
                      description: "Certificate issuer"
                      example: "CN=WR2,O=Google Trust Services,C=US"
                    cert_expires:
                      description: "Certificate expiry timestamp"
                      example: "2025-05-21T15:32:54.000Z"
                    cert_fingerprint:
                      type: string
                      description: "Certificate sha1 fingerprint"
                      example: "2f067cd1a347af920b67399349187666223253cb"